+++ PungDeong/PungDeong/Info.plist
@@ -9,7 +9,7 @@
 			<string>Editor</string>
 			<key>CFBundleURLSchemes</key>
 			<array>
-				<string>kakaod279a8725a4a2fcccbfcb999e0f80e14://kakaolink</string>
+				<string>kakaoc99afea96191236b7d142ebe483ca142://kakaolink</string>
 			</array>
 		</dict>
 	</array>
