+++ PungDeong/PungDeong/PungDeongApp.swift
@@ -6,12 +6,35 @@
 //
 
 import SwiftUI
+import KakaoSDKCommon
 
 @main
 struct PungDeongApp: App {
+    
+    //SwiftUI 지원 안되는 라이브러리 사용을 위해 선언
+    @UIApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
+    
+    
     var body: some Scene {
         WindowGroup {
             TestResultView()
         }
+    }
+}
+
+
+//MARK: - AppDelegate SwiftUI 지원 안되는 라이브러리 사용을 위해 선언
+class AppDelegate: NSObject, UIApplicationDelegate {
+    
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool
+    {
+                //패키지 초기화 작업 수행
+
+        KakaoSDK.initSDK(appKey: "c99afea96191236b7d142ebe483ca142")
+//        if let url = URLContexts.first?.url {
+//                    print("url : \(url)")
+//        }
+
+        return true
     }
 }
