+++ PungDeong/PungDeong/TestResult/TestResultViewModel.swift
@@ -5,16 +5,35 @@
 //  Created by GOngTAE on 2022/04/07.
 //
 
-import Foundation
-import Alamofire
-import KakaoSDKLink
+import SwiftUI
 
+import Alamofire
+
+import KakaoSDKLink
+import KakaoSDKCommon
 // 기본 템플릿 사용 시 필요
 import KakaoSDKTemplate
 
+
+
 class TestResultViewModel {
     
-    func shareKakao() {
+    static func shareKakao() {
         
+        //결과 공유  type 1 ID
+        let templateId = 74829
+        
+        LinkApi.shared.customLink(templateId: Int64(templateId)) { (linkResult, error) in
+            if let error = error {
+                print(error)
+            }
+            else {
+                print("customLink() success.")
+                if let linkResult = linkResult {
+                    // 카카오톡 오픈
+                    UIApplication.shared.open(linkResult.url, options: [:], completionHandler: nil)
+                }
+            }
+        }
     }
 }
